refactor: redesign ui
