+++ lang/zh.json
@@ -2,8 +2,10 @@
   "default_text_tips": "由 OpenAI 和 Vercel 提供支持\n- 点击下方发起聊天，`Enter`键发送，`Shift`+`Enter`键换行\n- 输入`/`或者点击左下角按钮添加预设提示",
   "default_image_tips": "由 OpenAI 和 Vercel 提供支持\n- 点击下方发起聊天，`Enter`键发送，`Shift`+`Enter`键换行\n- 图片生成不支持连续对话，直接输入你想要的图片效果，例如：`一只猫`\n- 图片链接的有效访问时间为 `2` 小时，如有需要请及时保存",
   "chat_placeholder": "开始聊天",
+  "search_placeholder": "搜索",
   "action_clear": "清空所有会话",
   "action_prompt": "添加预设提示",
+  "configuration": "设置",
   "config_model": "模型：",
   "config_save": "保留所有会话：",
   "config_continuous": "连续对话：",
@@ -11,8 +13,9 @@
   "config_temperature": "发散程度：",
   "config_images_count": "生成图片数：",
   "config_images_size": "图片尺寸：",
+  "config_language": "语言：",
   "status_loading": "AI 正在思考...",
-  "status_empty": "空对话",
+  "status_empty": "新的对话",
   "status_image_expired": "图片已过期",
   "chat_mode_text": "文本对话",
   "chat_mode_image": "图像生成对话"
