+++ package.json
@@ -26,6 +26,7 @@
     "@astrojs/vercel": "^3.2.1",
     "antd": "^5.3.2",
     "astro": "^2.1.3",
+    "dayjs": "^1.11.7",
     "eventsource-parser": "^1.0.0",
     "highlight.js": "^11.7.0",
     "katex": "^0.16.4",
