--- src/components/ConversationTabs/index.css
@@ -1,3 +1,0 @@
-.add-chat-icon .ant-dropdown-menu::-webkit-scrollbar {
-  display: none;
-}
