+++ src/interfaces/index.ts
@@ -36,3 +36,13 @@ export interface Prompt {
   act: string;
   prompt: string;
 }
+
+export interface RecordCardItem {
+  key: string;
+  title: string;
+  mode: ConversationMode;
+  message: string; // last message
+  time?: number; // last message time
+}
+
+export type ReactSetState<T> = React.Dispatch<React.SetStateAction<T>>;
