--- src/mock/markdown.ts
@@ -1,2 +1,0 @@
-export const mockMarkdownData =
-  "# markdown\n```typescript\n// markdown样式\nimport {marked} from 'marked' // 导入markdown转换器\nimport hljs from 'highlight.js' // 导入markdown代码块高亮包\n```";
