+++ src/modules/Content/MessageInput.tsx
@@ -8,16 +8,25 @@ import {
 } from 'react';
 import { Input } from 'antd';
 import GlobalContext from '@contexts/global';
-import PromptSelect from './PromptSelect';
+import PromptSelect from '@components/PromptSelect';
 
 const MessageInput: FC<{
   text: string;
   setText: (text: string) => void;
+  currentId: string; // current conversation id
   showPrompt: boolean;
   setShowPrompt: (showPrompt: boolean) => void;
   onSubmit: (message: string) => Promise<void>;
   loading: boolean;
-}> = ({ text, setText, showPrompt, setShowPrompt, onSubmit, loading }) => {
+}> = ({
+  text,
+  setText,
+  currentId,
+  showPrompt,
+  setShowPrompt,
+  onSubmit,
+  loading,
+}) => {
   const { i18n } = useContext(GlobalContext);
   const [promptKeyword, setPromptKeyword] = useState('');
   const [isInputComposition, setIsInputComposition] = useState(false);
@@ -49,41 +58,49 @@ const MessageInput: FC<{
     }
   }, [showPrompt]);
 
+  useEffect(() => {
+    // auto focus when change conversation
+    ref.current.focus();
+  }, [currentId]);
+
   return (
-    <div className="flex items-center">
+    <div className="flex items-center p-5 pt-0">
       <PromptSelect
         keyword={promptKeyword}
         showPrompt={showPrompt}
         onSelect={onPromptSelect}
       >
-        <Input.TextArea
-          ref={ref}
-          placeholder={i18n.chat_placeholder}
-          value={text}
-          autoFocus
-          onChange={(event) => {
-            const val = event.target.value;
-            setText(val);
-            if (val.startsWith('/')) {
-              setShowPrompt(true);
-              setPromptKeyword(val.slice(1));
-            } else {
-              setShowPrompt(false);
-              setPromptKeyword('');
-            }
-          }}
-          onCompositionStart={() => setIsInputComposition(true)}
-          onCompositionEnd={() => setIsInputComposition(false)}
-          onPressEnter={(e) => {
-            if (!e.shiftKey && !isInputComposition) {
-              e.preventDefault();
-              handleSubmit();
-            }
-          }}
-          size="large"
-          autoSize={{ minRows: 1, maxRows: 5 }}
-          allowClear
-        />
+        <div className="flex-1 border border-[#dfdfdf] rounded-lg ">
+          <Input.TextArea
+            ref={ref}
+            placeholder={i18n.chat_placeholder}
+            value={text}
+            autoFocus
+            bordered={false}
+            onChange={(event) => {
+              const val = event.target.value;
+              setText(val);
+              if (val.startsWith('/')) {
+                setShowPrompt(true);
+                setPromptKeyword(val.slice(1));
+              } else {
+                setShowPrompt(false);
+                setPromptKeyword('');
+              }
+            }}
+            onCompositionStart={() => setIsInputComposition(true)}
+            onCompositionEnd={() => setIsInputComposition(false)}
+            onPressEnter={(e) => {
+              if (!e.shiftKey && !isInputComposition) {
+                e.preventDefault();
+                handleSubmit();
+              }
+            }}
+            size="large"
+            autoSize={{ minRows: 1, maxRows: 5 }}
+            allowClear
+          />
+        </div>
       </PromptSelect>
 
       <i
