+++ src/modules/Sidebar/index.tsx
@@ -1,0 +1,120 @@
+import { FC, useContext } from 'react';
+import { Dropdown, Input, MenuProps } from 'antd';
+import omit from 'lodash.omit';
+import GlobalContext from '@contexts/global';
+import { getMaxIndex } from '@utils';
+import {
+  Conversation,
+  ConversationMode,
+  ReactSetState,
+  RecordCardItem,
+} from '@interfaces';
+import './index.css';
+import RecordCard from './RecordCard';
+
+const Sidebar: FC<{
+  data: RecordCardItem[];
+  setConversations: ReactSetState<Record<string, Conversation>>;
+  currentTab: string;
+  setCurrentTab: (tab: string) => void;
+}> = ({ data, setConversations, currentTab, setCurrentTab }) => {
+  const { i18n } = useContext(GlobalContext);
+
+  const onAdd = (mode: ConversationMode = 'text') => {
+    const id = getMaxIndex(data).toString();
+    setConversations((items) => ({
+      ...items,
+      [id]: {
+        id,
+        title: i18n.status_empty,
+        mode,
+        messages: [],
+        createdAt: Date.now(),
+      },
+    }));
+    setCurrentTab(id);
+  };
+
+  const onDelete = (key: string) => {
+    setConversations((items) => omit(items, [key]));
+    // delete other conversation doesnt need to update currentTab
+    if (currentTab === key) {
+      setCurrentTab(data.filter((tab) => tab.key !== key)[0]?.key);
+    }
+  };
+
+  const items: MenuProps['items'] = [
+    {
+      key: '1',
+      label: (
+        <div onClick={() => onAdd()}>
+          <i className="ri-chat-4-line align-bottom mr-1" />
+          {i18n.chat_mode_text}
+        </div>
+      ),
+    },
+    {
+      key: '2',
+      label: (
+        <div onClick={() => onAdd('image')}>
+          <i className="ri-image-line align-bottom mr-1" />
+          {i18n.chat_mode_image}
+        </div>
+      ),
+    },
+  ];
+
+  return (
+    <div className="flex flex-col h-full">
+      <div className="p-4 mt-2 flex items-baseline justify-between">
+        <span className="text-3xl text-gradient font-[800]">ChatGPT</span>
+        {/* <div className="text-sm ml-2">Powered by OpenAI and Vercel</div> */}
+        <a
+          href="https://github.com/GPTGenius/chatgpt-vercel"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <i className="ml-2 ri-github-fill text-xl" />
+        </a>
+      </div>
+      <div className="p-2 flex items-center justify-between mb-4">
+        <div className="rounded-xl h-10 border flex-1">
+          <Input
+            className="h-[100%]"
+            prefix={<i className="ri-search-2-line" />}
+            placeholder={i18n.search_placeholder}
+            bordered={false}
+            allowClear
+          />
+        </div>
+        <Dropdown
+          menu={{ items }}
+          trigger={['click']}
+          placement="bottomRight"
+          getPopupContainer={(node) => node}
+        >
+          <i className="ri-chat-new-line cursor-pointer p-2 ml-1" />
+        </Dropdown>
+      </div>
+      <div className="common-scrollbar flex-1 p-2 pt-0 overflow-auto">
+        {data.map((conversation, index) => (
+          <div key={conversation.key}>
+            {index !== 0 ? (
+              <div className="h-[1px] bg-[#edeeee] ml-2 mr-2" />
+            ) : null}
+            <RecordCard
+              data={conversation}
+              selected={conversation.key === currentTab}
+              onSelect={() => setCurrentTab(conversation.key)}
+              onDelete={
+                data.length > 1 ? () => onDelete(conversation.key) : null
+              }
+            />
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default Sidebar;
